Convert Data Structures to rang edgelist

This generic function converts several data structures provided by rang into an edgelist of package dependencies.

Usage

convert_edgelist(x, ...)

# Default S3 method
convert_edgelist(x, ...)

# S3 method for class 'ranglet'
convert_edgelist(x, ...)

# S3 method for class 'rang'
convert_edgelist(x, ...)

Arguments

x,

supported data structures are rang and ranglet S3 objects

...,

not used

Value

a data frame of directed edges of dependencies

Details

the resulting data frame can be converted to an igraph object for plotting and analysis via the function igraph::graph_from_data_frame()

Examples

# \donttest{
if (interactive()) {
    graph <- resolve(pkgs = c("openNLP", "LDAvis", "topicmodels", "quanteda"),
                snapshot_date = "2020-01-16")
                
    # dependency edgelist of a single package
    convert_edgelist(graph$ranglets[[1]])
    
    # full dependency edgelist
    convert_edgelist(graph)
}
# }