Export The Resolved Result As a renv Lockfile

This function exports the results from resolve() to a renv lockfile that can be used as an alternative to a docker container.

Usage

export_renv(rang, path = ".")

Arguments

rang

output from resolve()

path

character, path of the exported renv lockfile

Value

path, invisibly

Details

A renv lockfile is easier to handle than a docker container, but it cannot always reliably reproduce the exact computational environment,especially for very old code.

Examples

# \donttest{
if (interactive()) {
    graph <- resolve(pkgs = c("openNLP", "LDAvis", "topicmodels", "quanteda"),
                    snapshot_date = "2020-01-16")
    export_renv(graph, ".")
}
# }