Recipes for Building Container Images

A list containing several useful recipes for container building. Useful for the post_installation_steps argument of dockerize(). Available recipes are:

texlive: install pandoc and LaTeX, useful for rendering RMarkdown

texlivefull: Similar to the above, but install the full distribution of TeX Live (~ 3GB)

quarto: install quarto and tinytex

clean: clean up the container image by removing cache

make: install GNU make

Usage

recipes

Format

An object of class list of length 5.

Examples

# \donttest{
if (interactive()) {
    graph <- resolve(pkgs = c("openNLP", "LDAvis", "topicmodels", "quanteda"),
                    snapshot_date = "2020-01-16")
    ## install texlive
    dockerize(graph, ".", post_installation_steps = recipes[['texlive']])
}
# }